+++ packages/terra-core-docs/CHANGELOG.md
@@ -13,6 +13,7 @@
   * Migrate docs from terra-form-radio.
   * Migrate docs from terra-form-input.
   * Migrate docs from terra-form-textarea.
+  * Migrate docs from terra-grid.
 
 ## 1.12.0 - (November 2, 2021)
