+++ packages/terra-core-docs/src/terra-dev-site/doc/grid/Grid.1.doc.mdx
@@ -1,8 +1,8 @@
 import { Badge } from 'terra-grid/package.json?dev-site-package';
 
-import GridDefault from '../example/GridDefault?dev-site-example';
-import GridResponsive from '../example/GridResponsive?dev-site-example';
-import GridNested from '../example/GridNested?dev-site-example';
+import GridDefault from './example/GridDefault?dev-site-example';
+import GridResponsive from './example/GridResponsive?dev-site-example';
+import GridNested from './example/GridNested?dev-site-example';
 
 import GridPropsTable from 'terra-grid/src/Grid?dev-site-props-table';
 import GridColumnPropsTable from 'terra-grid/src/GridColumn?dev-site-props-table';
