+++ packages/terra-core-docs/src/terra-dev-site/doc/grid/example/GridResponsive.jsx
