+++ packages/terra-core-docs/src/terra-dev-site/test/grid/GridEnormous.test.jsx
@@ -1,70 +1,70 @@
 import React from 'react';
-import Grid from '../../../../lib/Grid';
+import Grid from 'terra-grid';
 import './GridExample.module.scss';
 
-const GridHuge = () => (
+const GridEnormous = () => (
   <Grid id="grid">
     <Grid.Row>
-      <Grid.Column id="column-1" huge={1}>
+      <Grid.Column id="column-1" enormous={1}>
         <div className="terra-grid-example">1 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-2" huge={2}>
+      <Grid.Column id="column-2" enormous={2}>
         <div className="terra-grid-example">2 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-3" huge={3}>
+      <Grid.Column id="column-3" enormous={3}>
         <div className="terra-grid-example">3 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-4" huge={4}>
+      <Grid.Column id="column-4" enormous={4}>
         <div className="terra-grid-example">4 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-5" huge={5}>
+      <Grid.Column id="column-5" enormous={5}>
         <div className="terra-grid-example">5 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-6" huge={6}>
+      <Grid.Column id="column-6" enormous={6}>
         <div className="terra-grid-example">6 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-7" huge={7}>
+      <Grid.Column id="column-7" enormous={7}>
         <div className="terra-grid-example">7 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-8" huge={8}>
+      <Grid.Column id="column-8" enormous={8}>
         <div className="terra-grid-example">8 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-9" huge={9}>
+      <Grid.Column id="column-9" enormous={9}>
         <div className="terra-grid-example">9 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-10" huge={10}>
+      <Grid.Column id="column-10" enormous={10}>
         <div className="terra-grid-example">10 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-11" huge={11}>
+      <Grid.Column id="column-11" enormous={11}>
         <div className="terra-grid-example">11 Column Span</div>
       </Grid.Column>
     </Grid.Row>
     <Grid.Row>
-      <Grid.Column id="column-12" huge={12}>
+      <Grid.Column id="column-12" enormous={12}>
         <div className="terra-grid-example">12 Column Span</div>
       </Grid.Column>
     </Grid.Row>
   </Grid>
 );
 
-export default GridHuge;
+export default GridEnormous;
